apps/web/pyproject.toml
[project]
name = "monorepo-web"
version = "0.1.0"
description = "Web application"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.116.0",
    "pydantic-settings>=2.10.0",
    "structlog>=25.4.0",
    "uvicorn>=0.35.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "authlib>=1.4.0",
    "redis[asyncio]>=5.0.0",
    "itsdangerous>=2.2.0",
    "httpx>=0.28.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.0",
    "taskipy>=1.14.1",
    "pytest>=8.4.0",
    "pytest-asyncio>=0.24.0",
    "fakeredis>=2.26.0",
    "ruff>=0.15.9",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.taskipy.tasks]
test = "python -m pytest tests/"
test-project = "python -m pytest tests/test_project_web.py"
